requirements.txt
langchain
langfuse
openai
python-dotenv
requests
langchain-openai
